fix review issues
