+++ client/starwhale/utils/config.py
@@ -1,13 +1,12 @@
 import os
+import sys
 import typing as t
 import getpass
-import tempfile
 import subprocess
-from sys import platform
+from shutil import which
 from pathlib import Path
 
 import yaml
-import click
 
 from starwhale.utils import load_yaml
 from starwhale.consts import (
@@ -23,9 +22,8 @@ from starwhale.consts import (
     OBJECT_STORE_DIRNAME,
     DEFAULT_SW_LOCAL_STORAGE,
 )
-from starwhale.utils.cli import AliasedGroup
 from starwhale.consts.env import SWEnv
-from starwhale.utils.error import ExistedError, NotFoundError, NoSupportError
+from starwhale.utils.error import NotFoundError, NoSupportError
 
 from . import console, now_str, fmt_http_server
 from .fs import ensure_dir, ensure_file
@@ -259,40 +257,10 @@ class SWCliConfigMixed:
         update_swcli_config(**self._config)
 
 
-@click.group(
-    "config",
-    cls=AliasedGroup,
-    help="Configuration management, edit is supported now",
-)
-def config_cmd() -> None:
-    pass
-
-
-@config_cmd.command("edit", aliases=["e"], help="edit the configuration of swlci")
-def __edit() -> None:
-    _edit()
-
-
-def _edit() -> None:
-    _editor = ""
-    if platform == "linux" or platform == "linux2":
-        _editor = os.environ.get(SWEnv.sw_editor, "")
-        if not _editor:
-            _editor = os.environ.get("EDITOR", "vi")
-    elif platform == "darwin" or platform == "win32":
-        _editor = os.environ.get(SWEnv.sw_editor, "")
-        if not _editor:
-            raise NoSupportError(f"No {SWEnv.sw_editor} is configured")
-    else:
-        raise NoSupportError(f"Current platform is not supported by: {platform}")
-    path_config = get_swcli_config_path()
-    fd, fname = tempfile.mkstemp()
-    with open(path_config) as config_f, os.fdopen(fd, "w") as tmpf:
-        for line in config_f:
-            tmpf.write(line)
-        tmpf.close()
-        cmd = _editor + " " + fname
-        _c = subprocess.call(cmd, shell=True)
-        if _c:
-            raise ExistedError("editing file with unexpected failure")
-        os.rename(fname, path_config)
+def edit_from_shell() -> None:
+    _editor = os.environ.get("EDITOR") or os.environ.get("VISUAL") or "vi"
+    if which(_editor.split()[0]) is None:
+        raise NoSupportError(
+            f"no found {_editor} bin in {sys.platform}. Please configure one using EDITOR or VISUAL environment variable"
+        )
+    subprocess.call(f"{_editor} {get_swcli_config_path()}", shell=True)
